packaged/hsv_test/requirements.txt
numpy
protobuf
matplotlib
scikit-learn
python-dotenv
gunicorn
pyyaml